+++ src/helper.hpp
@@ -13,9 +13,6 @@ class Helper
     static const QList<qint32> blueCodedList;
 
     public:
-        //enum LogTypes{ ADMIN = 0, COMMENT, USAGE, UPNP, BAN, DC, MUTE,
-        //               IGNORE = 7 };
-
         static QInputDialog* createInputDialog(QWidget* parent, QString& label,
                                                QInputDialog::InputMode mode,
                                                int width, int height);
@@ -44,10 +41,6 @@ class Helper
         static qint32 serNumtoInt(QString& sernum);
 
         static bool isBlueCodedSerNum(const quint32& sernum);
-
-//        static void logToFile(const LogTypes& type, const QString& text,
-//                              const bool& timeStamp = false,
-//                              const bool& newLine = false);
 
         static bool confirmAction(QWidget* parent, QString& title,
                                   QString& prompt);